add git safe directory in action
The github action seems to be broken.  The formatter runs successfully,
but git complains that it is not in a repo when checking if files have
changed.

This seems to be because of how directories are nested in the docker
environment.

Explicitly mark the directory as safe so git agrees to run.